refactor: Address code review feedback for CRM feature
I've incorporated feedback from the code review on the new client management feature.

The key improvements are:

1.  **More Robust Client Lookup:** The client matching logic in the backend (`convex/models/assessments.ts`) has been enhanced. It now prioritizes searching for clients by their email address before falling back to searching by name, significantly reducing the risk of creating duplicate records.

2.  **Refactored Client Selector UI:** The complex client search and selection UI has been extracted from the main assessment form into its own reusable component (`app/(client)/assessment/new/ClientSelector.tsx`). This greatly improves the readability and maintainability of the main form page.

3.  **Enhanced UI Feedback:** Your experience on the assessment form has been improved by adding clear feedback for asynchronous operations. The form now displays loading indicators, disables the main button during submission, and uses toast notifications to clearly communicate success or failure to you.

**Note on Verification:**
I was unable to run the development server and build scripts due to persistent configuration issues. As a result, I could not complete the end-to-end and visual verification.
