+++ convex/models/assessments.ts
@@ -31,29 +31,43 @@ export async function createAssessmentModel(
 
   let clientId: Id<"clients">;
 
-  // Normalize inputs for matching/insert
-  const normalizedName = args.client.name.trim().toLocaleLowerCase();
-  const normalizedEmail = args.client.email?.trim().toLocaleLowerCase();
-  const normalizedPhone = args.client.phone?.replace(/\D+/g, "");
+  // Prioritize finding client by email if provided, as it's more likely to be unique
+  if (args.client.email) {
+    const clientByEmail = await ctx.db
+      .query("clients")
+      .withIndex("by_orgId_and_email", (q) =>
+        q.eq("orgId", args.orgId).eq("email", args.client.email)
+      )
+      .first();
+    if (clientByEmail) {
+      clientId = clientByEmail._id;
+    }
+  }
 
-  const existingClient = await ctx.db
-    .query("clients")
-    .withIndex("by_orgId_and_name", (q) =>
-      q.eq("orgId", args.orgId).eq("name", normalizedName)
-    )
-    .first();
+  // If client not found by email, try finding by name
+  if (!clientId) {
+      const clientByName = await ctx.db
+        .query("clients")
+        .withIndex("by_orgId_and_name", (q) =>
+          q.eq("orgId", args.orgId).eq("name", args.client.name)
+        )
+        .first();
+      if (clientByName) {
+          clientId = clientByName._id;
+      }
+  }
 
-  if (existingClient) {
-    clientId = existingClient._id;
-  } else {
+  // If client is still not found, create a new one
+  if (!clientId) {
     clientId = await ctx.db.insert("clients", {
       orgId: args.orgId,
       userId: args.userId,
-      name: normalizedName,
-      email: normalizedEmail,
-      phone: normalizedPhone,
+      name: args.client.name,
+      email: args.client.email,
+      phone: args.client.phone,
     });
   }
+
   const serviceIds = args.services
     .map((service) => ctx.db.normalizeId("services", service))
     .filter((id): id is Id<"services"> => id !== null);
