+++ convex/schema.ts
@@ -150,6 +150,7 @@ export default defineSchema({
   }).index("by_orgId", ["orgId"])
     .index("by_name", ["name"])
     .index("by_orgId_and_name", ["orgId", "name"])
+    .index("by_orgId_and_email", ["orgId", "email"])
     .searchIndex("search_name", {
       searchField: "name",
       filterFields: ["orgId"],
